feat: add script to provision Jenkins environment secrets and start service

+++ 00_Setup/Infra/jenkins/env_setup/env_setup.sh
@@ -2,10 +2,23 @@
 
 set -euo pipefail
 
+ENV_FILE="./jenkins.env"
+
+if [[ ! -f "$ENV_FILE" ]]; then
+    echo "ERROR: $ENV_FILE not found"
+    exit 1
+fi
+
+set -a
+source "$ENV_FILE"
+set +a
+
 echo "Create the env file"
+
 sudo touch /etc/jenkins/jenkins-secrets.env
 
 echo "Add the ENVs"
+
 sudo tee /etc/jenkins/jenkins-secrets.env > /dev/null <<EOF
 JENKINS_ADMIN_PASSWORD=${JENKINS_ADMIN_PASSWORD}
 JENKINS_USER1_PASSWORD=${JENKINS_USER1_PASSWORD}
@@ -18,13 +31,12 @@ COSIGN_PRIVATE_KEY=${COSIGN_PRIVATE_KEY}
 COSIGN_KEY_PASSWORD=${COSIGN_KEY_PASSWORD}
 EOF
 
-echo "Secure the file by modifying users and permissions"
 sudo chown root:root /etc/jenkins/jenkins-secrets.env
 sudo chmod 600 /etc/jenkins/jenkins-secrets.env
 
-echo "Reload Daemon, Start and Enable Jenkins, Checking the Status"
 sudo systemctl daemon-reload
 sudo systemctl enable jenkins
 sudo systemctl start jenkins
-sudo systemctl status Jenkins --no-pager
-sudo journalctl -u Jenkins -f
+
+sudo systemctl status jenkins --no-pager
+sudo journalctl -u jenkins -f
